requirements.txt
torch==2.9.0
torchvision==0.24.0
torchaudio==2.9.0
soundfile
numpy
torch_audiomentations
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core
wget
gdown
datasets
yadisk
torchcodec
librosa
speechbrain
git+https://github.com/AndreevP/wvmos

black
isort
pre-commit
